Removed ink, no-undo performers
Under the new Material/BlockData convention, Ink performers will be useless. This is because BlockData contains a reference to Material, and the data is named and block-specific. Thus, when you want to set BlockData, you will have to change the entire BlockData, including the Material that comes along with it. This renders the Ink performer as useless, because changing BlockData with Material would just mean that it would be the Combo performer.

Also, why do we have no-undo performers? Dangerous!